set up page structure at top level
