+++ src/tapis-app/Router/Router.tsx
@@ -1,25 +1,14 @@
-import { hot } from 'react-hot-loader/root';
 import React from 'react';
-import { Redirect, Route } from 'react-router-dom';
-import { Apps, Login, Dashboard, Jobs, Systems } from 'tapis-app/Sections';
-import { Layout } from 'tapis-app/Layout';
+import { Redirect, Route, Switch } from 'react-router-dom';
+import { Apps, Login, Dashboard, Jobs, Systems, UIPatterns } from 'tapis-app/pages';
 import { SectionHeader, ProtectedRoute } from 'tapis-ui/_common';
 import { useLogin } from 'tapis-hooks/authenticator';
-import Sidebar from '../Sidebar/Sidebar';
-import UIPatterns from '../UIPatterns';
-import './App.scss';
 
-const App: React.FC = () => {
+const AppRouter: React.FC = () => {
   const { logout } = useLogin();
 
-  const header = (
-    <div>
-      <SectionHeader className="tapis-ui__header">TAPIS UI</SectionHeader>
-    </div>
-  );
-
-  const workbenchContent = (
-    <div className="workbench-content">
+  return (
+    <Switch>
       <Route exact path="/">
         <Dashboard />
       </Route>
@@ -46,14 +35,8 @@ const App: React.FC = () => {
         <SectionHeader>UI Patterns</SectionHeader>
         <UIPatterns />
       </Route>
-    </div>
-  );
-
-  return (
-    <div style={{ display: 'flex', flexGrow: 1, height: '100vh' }}>
-      <Layout top={header} left={<Sidebar />} right={workbenchContent} />
-    </div>
+    </Switch>
   );
 };
 
-export default hot(App);
+export default AppRouter;
