+++ src/tapis-app/Router/index.ts
@@ -1,0 +1,1 @@
+export { default as Router } from './Router';
