+++ src/tapis-app/components/Layout/Layout.tsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import styles from './Layout.module.scss';
+
+interface LayoutProps {
+  left?: React.ReactNode;
+  right?: React.ReactNode;
+  top?: React.ReactNode;
+  bottom?: React.ReactNode;
+}
+const Layout: React.FC<LayoutProps> = ({ left, right, top, bottom }) => (
+  <div className={styles['layout-root']}>
+    {top}
+    <div className={styles['layout-row']}>
+      {left}
+      {right}
+    </div>
+    {bottom}
+  </div>
+);
+
+export default Layout;
