+++ src/tapis-app/components/Sidebar/Sidebar.module.scss
