+++ src/tapis-app/components/Sidebar/Sidebar.tsx
