+++ src/tapis-app/components/index.ts
@@ -1,0 +1,3 @@
+export { default as Layout } from './Layout';
+export { default as Sidebar } from './Sidebar';
+export { SectionHeader, SectionNavWrapper, SectionBody } from './Section';
