+++ src/tapis-app/pages/Dashboard/Dashboard.scss
