+++ src/tapis-app/pages/UIPatterns/UIPatterns.js
