+++ src/tapis-app/pages/UIPatterns/UIPatterns.module.scss
