+++ src/tapis-app/pages/UIPatterns/UIPatternsMessage/UIPatternsMessage.js
