+++ src/tapis-app/pages/UIPatterns/UIPatternsPill/UIPatternsPill.js
