+++ src/tapis-app/pages/UIPatterns/index.js
