+++ src/tapis-app/pages/index.ts
@@ -3,3 +3,4 @@ export { default as Dashboard } from './Dashboard';
 export { default as Jobs } from './Jobs';
 export { default as Login } from './Login';
 export { default as Systems } from './Systems';
+export { default as UIPatterns } from './UIPatterns';
